Move standardize method to utils file
